Extract raw query compiler module
